pyproject.toml
[project]
name = "gazelle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipywidgets>=8.1.5",
    "matplotlib>=3.10.0",
    "numpy>=2.2.0",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "retinaface-pytorch>=0.0.7",
    "scikit-learn>=1.6.0",
    "scipy>=1.14.1",
    "screeninfo>=0.8.1",
    "timm>=1.0.12",
    "torch==2.4.1",
    "torchaudio==2.4.1",
    "torchvision==0.19.1",
    "triton>=3.0.0",
    "xformers>=0.0.28.post1",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.3.4",
    "ruff>=0.8.3",
    "taskipy>=1.14.1",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "platform_system != 'Darwin'"},
]
torchvision = [
  { index = "pytorch-cu124", marker = "platform_system != 'Darwin'"},
]
torchaudio = [
  { index = "pytorch-cu124", marker = "platform_system != 'Darwin'" },
]
xformers = [
  { index = "pytorch-cu124", marker = "platform_system != 'Darwin'" },
]
triton = [
  { url = "https://huggingface.co/madbuda/triton-windows-builds/resolve/main/triton-3.0.0-cp312-cp312-win_amd64.whl", marker = "platform_system == 'Windows'" },
]

[tool.taskipy.tasks]
lint = "ruff check"
fmt = "ruff format"

[tool.ruff.format]
quote-style = "single"
